Display milliseconds up to two decimal places

+++ src/payload.rs
@@ -2,6 +2,6 @@ use super::*;
 
 #[derive(Debug, Clone, Serialize)]
 pub(crate) struct Payload {
-  pub(crate) time: u128,
+  pub(crate) time: f64,
   pub(crate) courses: Vec<Course>,
 }
